requirements.txt
allure-pytest==2.8.22
allure-python-commons==2.8.22
asttokens==2.0.3
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
distconfig3==1.0.1
idna==2.8
iniconfig==1.1.1
more-itertools==8.6.0
packaging==20.7
pathtools==0.1.2
pluggy==0.13.1
py==1.9.0
pycryptodome==3.9.9
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.1.2
PyYAML==5.3.1
requests==2.25.0
semantic-version==2.8.5
six==1.15.0
toml==0.10.2
urllib3==1.24.3
vyper-config==1.0.0
watchdog==0.10.4
